+++ .circleci/config.yml
@@ -153,7 +153,7 @@ jobs:
                   ddev xdebug on
                   $CIRCLE_WORKING_DIRECTORY/cc-test-reporter before-build
                   ddev dkan-phpunit \
-                    --group dataset \
+                    --group functional \
                     --coverage-clover /var/www/html/docroot/modules/contrib/dkan/clover.xml \
                     --coverage-html /var/www/html/docroot/modules/contrib/dkan/coverage-html \
                     --log-junit /var/www/html/docroot/modules/contrib/dkan/junit/junit.xml
@@ -178,7 +178,7 @@ jobs:
                 name: Run PHPUnit tests
                 command: |
                   ddev dkan-phpunit \
-                    --group dataset \
+                    --group functional \
                     --log-junit /var/www/html/docroot/modules/contrib/dkan/junit/junit.xml
       - store_test_results:
           path: dkan/junit
