+++ modules/datastore/tests/src/Functional/DictionaryEnforcerTest.php
@@ -18,7 +18,7 @@ use weitzman\DrupalTestTraits\ExistingSiteBase;
  *
  * @package Drupal\Tests\datastore\Functional
  * @group datastore
- * @group _dataset
+ * @group functional
  */
 class DictionaryEnforcerTest extends ExistingSiteBase {
