+++ modules/metastore/tests/src/Functional/Api1/DatasetRevisionTest.php
@@ -6,7 +6,7 @@ use Drupal\Tests\common\Functional\Api1TestBase;
 use GuzzleHttp\RequestOptions;
 
 /**
- * @group dataset
+ * @group functional
  */
 class DatasetRevisionTest extends Api1TestBase {
