+++ modules/metastore/tests/src/Functional/Api1/SchemaTest.php
@@ -5,7 +5,7 @@ namespace Drupal\Tests\metastore\Functional\Api1;
 use Drupal\Tests\common\Functional\Api1TestBase;
 
 /**
- * @group dataset
+ * @group functional
  */
 class SchemaTest extends Api1TestBase {
