+++ modules/metastore/tests/src/Functional/OrphanCheckerTest.php
@@ -13,7 +13,7 @@ use weitzman\DrupalTestTraits\ExistingSiteBase;
  *
  * @package Drupal\Tests\metastore\Functional
  * @group metastore
- * @group dataset
+ * @group functional
  */
 class OrphanCheckerTest extends ExistingSiteBase {
   use GetDataTrait;
