+++ modules/metastore/tests/src/Functional/Storage/NodeDataTest.php
@@ -15,7 +15,7 @@ use weitzman\DrupalTestTraits\ExistingSiteBase;
  *
  * @package Drupal\Tests\dkan\Functional
  * @group dkan
- * @group dataset
+ * @group functional
  */
 class NodeDataTest extends ExistingSiteBase {
   use CleanUp;
